requirements.txt
torch==2.0.1
accelerate==0.21.0
bitsandbytes==0.41.0
deepspeed==0.9.5
einops-exts==0.0.4
einops==0.6.1
gradio==3.32.0
gradio_client==0.2.9
httpx==0.24.0
markdown2==2.4.10
numpy==1.26.0
peft==0.4.0
scikit-learn==1.2.2
sentencepiece
shortuuid==1.0.11
timm==0.6.13
torchvision==0.15.2
transformers==4.31.0
wandb==0.15.12
wavedrom==2.0.3.post3
Pygments==2.16.1
omegaconf
pytorch_lightning==2.1
scikit-image
opencv-python-headless
lpips
scipy
tiktoken
safetensors
transformers_stream_generator
tokenizers==0.13.3
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0